requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
certifi==2025.8.3
deprecation==2.1.0
django==5.2.5
django-filter==25.1
djangorestframework==3.16.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
markdown==3.8.2
packaging==25.0
postgrest==1.1.1
pydantic==2.11.7
pydantic-core==2.33.2
pyjwt==2.10.1
python-dateutil==2.9.0.post0
realtime==2.7.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
storage3==0.12.1
strenum==0.4.15
supabase==2.18.1
supabase-auth==2.12.3
supabase-functions==0.10.1
typing-extensions==4.14.1
typing-inspection==0.4.1
tzdata==2025.2
websockets==15.0.1
